#27 split into multiple modules for easier importing in external projects
