+++ fs-core/src/main/java/be/mathiasbosman/fs/core/config/Generated.java
@@ -1,0 +1,17 @@
+package be.mathiasbosman.fs.core.config;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Annotation for generated methods. Can be used to ignore them in test coverage
+ */
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface Generated {
+
+}
