+++ fs-core/src/main/java/be/mathiasbosman/fs/core/domain/FileSystemNodeType.java
@@ -1,0 +1,11 @@
+package be.mathiasbosman.fs.core.domain;
+
+/**
+ * Types for {@link FileSystemNode}.
+ *
+ * @author mathiasbosman
+ * @since 0.0.1
+ */
+public enum FileSystemNodeType {
+  FILE, DIRECTORY, NONE_EXISTENT
+}
