+++ fs-nextcloud/pom.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xmlns="http://maven.apache.org/POM/4.0.0"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <artifactId>fs-nextcloud</artifactId>
+  <dependencies>
+    <dependency>
+      <artifactId>fs-core</artifactId>
+      <groupId>be.mathiasbosman</groupId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <artifactId>fs-test</artifactId>
+      <groupId>be.mathiasbosman</groupId>
+      <scope>test</scope>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <artifactId>nextcloud-api</artifactId>
+      <groupId>org.aarboard.nextcloud</groupId>
+    </dependency>
+    <dependency>
+      <artifactId>slf4j-api</artifactId>
+      <groupId>org.slf4j</groupId>
+    </dependency>
+    <dependency>
+      <artifactId>spring-boot-starter-test</artifactId>
+      <groupId>org.springframework.boot</groupId>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
+
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <artifactId>file-services</artifactId>
+    <groupId>be.mathiasbosman</groupId>
+    <version>1.0.7-SNAPSHOT</version>
+  </parent>
+
+</project>
