+++ fs-s3/src/test/resources/docker/docker-compose-minio.yml
