+++ fs-s3/src/test/resources/logback-test.xml
@@ -1,0 +1,15 @@
+<configuration>
+  <appender class="ch.qos.logback.core.ConsoleAppender" name="STDOUT">
+    <encoder>
+      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger - %msg%n</pattern>
+    </encoder>
+  </appender>
+
+  <logger level="INFO" name="be.mathiasbosman"/>
+
+  <logger level="INFO" name="org.testcontainers"/>
+  <logger level="WARN" name="com.github.dockerjava"/>
+  <root level="info">
+    <appender-ref ref="STDOUT"/>
+  </root>
+</configuration>
