+++ fs-test/src/main/java/be/mathiasbosman/fs/test/ContainerServiceDto.java
@@ -1,0 +1,15 @@
+package be.mathiasbosman.fs.test;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * Dto to hold container information.
+ */
+@Getter
+@AllArgsConstructor
+public class ContainerServiceDto {
+
+  private final String serviceName;
+  private final int port;
+}
