+++ pom.xml
@@ -3,6 +3,46 @@
   xmlns="http://maven.apache.org/POM/4.0.0"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <artifactId>file-services</artifactId>
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <artifactId>guava</artifactId>
+        <groupId>com.google.guava</groupId>
+        <version>30.1.1-jre</version>
+      </dependency>
+      <dependency>
+        <artifactId>commons-io</artifactId>
+        <groupId>commons-io</groupId>
+        <version>2.11.0</version>
+      </dependency>
+      <!-- file services -->
+      <dependency>
+        <artifactId>aws-java-sdk-s3</artifactId>
+        <groupId>com.amazonaws</groupId>
+        <version>1.12.61</version>
+      </dependency>
+      <dependency>
+        <artifactId>nextcloud-api</artifactId>
+        <groupId>org.aarboard.nextcloud</groupId>
+        <version>11.5.0</version>
+      </dependency>
+      <dependency>
+        <artifactId>testcontainers</artifactId>
+        <groupId>org.testcontainers</groupId>
+        <version>1.16.0</version>
+      </dependency>
+      <dependency>
+        <artifactId>commons-lang3</artifactId>
+        <groupId>org.apache.commons</groupId>
+        <version>3.12.0</version>
+      </dependency>
+      <dependency>
+        <artifactId>commons-collections4</artifactId>
+        <groupId>org.apache.commons</groupId>
+        <version>4.4</version>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
 
   <build>
     <plugins>
@@ -31,57 +71,15 @@
       </plugin>
     </plugins>
   </build>
+  <modules>
+    <module>fs-core</module>
+    <module>fs-s3</module>
+    <module>fs-test</module>
+    <module>fs-nextcloud</module>
+    <module>fs-nio</module>
+  </modules>
 
-  <dependencies>
-    <dependency>
-      <artifactId>commons-lang3</artifactId>
-      <groupId>org.apache.commons</groupId>
-      <version>3.12.0</version>
-    </dependency>
-    <dependency>
-      <artifactId>commons-collections4</artifactId>
-      <groupId>org.apache.commons</groupId>
-      <version>4.4</version>
-    </dependency>
-    <dependency>
-      <artifactId>guava</artifactId>
-      <groupId>com.google.guava</groupId>
-      <version>30.1.1-jre</version>
-    </dependency>
-    <dependency>
-      <artifactId>commons-io</artifactId>
-      <groupId>commons-io</groupId>
-      <version>2.11.0</version>
-    </dependency>
-    <dependency>
-      <artifactId>lombok</artifactId>
-      <groupId>org.projectlombok</groupId>
-    </dependency>
-    <!-- file services -->
-    <dependency>
-      <artifactId>aws-java-sdk-s3</artifactId>
-      <groupId>com.amazonaws</groupId>
-      <version>1.12.61</version>
-    </dependency>
-    <dependency>
-      <groupId>org.aarboard.nextcloud</groupId>
-      <artifactId>nextcloud-api</artifactId>
-      <version>11.5.0</version>
-    </dependency>
-
-    <!-- testing -->
-    <dependency>
-      <artifactId>testcontainers</artifactId>
-      <groupId>org.testcontainers</groupId>
-      <scope>test</scope>
-      <version>1.16.0</version>
-    </dependency>
-    <dependency>
-      <artifactId>spring-boot-starter-test</artifactId>
-      <groupId>org.springframework.boot</groupId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
+  <packaging>pom</packaging>
   <description>File services</description>
   <distributionManagement>
     <!-- Github distribution -->
