--- src/main/java/be/mathiasbosman/fs/config/Generated.java
@@ -1,17 +1,0 @@
-package be.mathiasbosman.fs.config;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Annotation for generated methods. Can be used to ignore them in test coverage
- */
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE, ElementType.METHOD})
-public @interface Generated {
-
-}
