--- src/main/java/be/mathiasbosman/fs/domain/FileSystemNodeType.java
@@ -1,11 +1,0 @@
-package be.mathiasbosman.fs.domain;
-
-/**
- * Types for {@link FileSystemNode}.
- *
- * @author mathiasbosman
- * @since 0.0.1
- */
-public enum FileSystemNodeType {
-  FILE, DIRECTORY, NONE_EXISTENT
-}
