--- src/test/java/be/mathiasbosman/fs/ContainerServiceDto.java
@@ -1,12 +1,0 @@
-package be.mathiasbosman.fs;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ContainerServiceDto {
-
-  private final String serviceName;
-  private final int port;
-}
